As part of decoupling our microservices we need to start to use domain objects correctly. 'StageWithCaseData' is not part of the domain, any call that currently returns a 'StageWithCaseData' is a smell and should return a Case instead. A 'stage with a case data' is actually a Case as case contains stage.
This PR is a refactor to create a SearchService which initially retrieves data from StageService and later will be changed to use CaseService  without having to change the rest of the code. This code has moved from StageService but no functionality has been changed.
